Updates for account controller

+++ database/migrations/2023_05_03_023956_add_shopify_user_id.php
@@ -28,14 +28,13 @@ return new class extends Migration
         });
 
         Schema::table('users', function (Illuminate\Database\Schema\Blueprint $table) {
-            $table->unsignedInteger('shopify_user_id')->index()->nullable();
+            $table->unsignedBigInteger('shopify_user_id')->index()->nullable();
         });
 
         Schema::table('companies', function(Illuminate\Database\Schema\Blueprint $table){
                 $table->string('shopify_name')->index()->nullable();
                 $table->string('shopify_access_token')->index()->nullable();
         });
-
 
     }
